+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added config option to look for wrappers recursively
 - Exported `getCompilerConfigForContract` function for plugin support
 - Added request timout configuration
 - Added docs for script args